moving more stuff to FitBase
